requirements.txt
google-cloud-bigquery
db-dtypes
google-cloud-secret-manager
